docqa-proto/requirements.txt
openai>=1.51.0
tiktoken>=0.7.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
python-docx>=1.1.2
unstructured>=0.14.6  # optional; we’ll keep our own loaders simple
pydantic>=2.8.2
fastapi>=0.115.0
uvicorn>=0.30.3
sqlalchemy>=2.0.35
